Move some stuff around and clean-up imports
