+++ corehq/apps/hqadmin/templates/hqadmin/system_info.html
@@ -408,7 +408,7 @@
                             <span class="label label-danger">DOWN</span>
                         {% endif %}
                         <hr>
-                        {{ celery.msg|safe }}<br>
+                        {{ celery_stats|safe }}<br>
                     </td>
                     <td>
                         <dl class="dl-horizontal">
